fs: Provide _remove_dir_contents
This splits remove_dir_all into two: one part is the thin wrapper for
_delete_dir_contents, and the other part is responsible for deleting
the top-level directory.

We are going to expose a function for only deleting the contents, and
this will be its primary building block.

I provided two entrypoints, one of which wraps the other, rather than
a boolean and a conditional, because IMO it mekes the code clearer.
(It would also have been possible to duplicate the canonicalize call,
but I felt that was worse.)

There is no functional change yet.

Signed-off-by: Ian Jackson <ijackson@chiark.greenend.org.uk>