unix: Provide _remove_dir_contents
The implementation is relatively straightforward.  It's a shame that
we need to reach for libc to get the error handling right!

The return type is different to that of Windows's fs:_remove_dir_all.
Unix does not need to canonicalize the path, so doesn't return
a canonicalised path.

This type difference is OK provided both versions typecheck at the
call site, which will appear in a moment.

Right now this code is not used - the caller will appear in the next
commit - so we must temporarily add `#[allow]`.

As discussed, and differently to my v1,
  https://github.com/XAMPPRocky/remove_dir_all/pull/22/files/30791c382ca9535d4fd9758e1b070393bfa6ef3e#diff-cff001b3dd9b58fd8acadc50ffc32f83
I have made remove_dir_contents intolerant of ENOENT.

Signed-off-by: Ian Jackson <ijackson@chiark.greenend.org.uk>