Remove unnecessary code
The last case would always return None since we already checked whether
this is a directory. In case we were handling a file and no error (with
the exception of is_eloop) occurred, set the file handle to None.
For clarity, the file handle variable has been renamed to
child_directory.